+++ stack.h
@@ -15,16 +15,16 @@
 struct stack {
     uint32_t maxsize;
     int32_t top;
-    int32_t *data;
+    float *data;
 };
 
 struct stack *init_stack(uint32_t size);
 uint32_t stack_size(struct stack *st);
 int8_t stack_isempty(struct stack *st);
 int8_t stack_isfull(struct stack *st);
-int8_t push(struct stack *st, int32_t val);
-int8_t pop(struct stack *st, int32_t *popval);
-int32_t peek(struct stack *st);
+int8_t push(struct stack *st, float val);
+int8_t pop(struct stack *st, float *popval);
+float peek(struct stack *st);
 void free_stack(struct stack *st);
 
 #endif //STACK_H
